\name{atomic}
\alias{atomic}
\title{
Atomic random variables
}
\description{
In contrast to \link{derived} random variables, atomic random variables (RVs) do not depend on
other RVs and are mutually independent from all other atomic RVs. These RVs form the basis from
which a complex system of dependent RVs can be assembled. Like all RVs, atomic RVs are constructed
using factory functions as those listed below.

}
\seealso{
 \code{\link{delta}},
 \code{\link{uniform}},
 \code{\link{normal}},
 \code{\link{gamma}},
 \code{\link{invgamma}},
 \code{\link{weibull}}.
}
