\name{stochbb-package}
\alias{stochbb-package}
\alias{stochbb}
\docType{package}
\title{
\packageTitle{stochbb}
}
\description{
\packageDescription{stochbb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stochbb}
\packageIndices{stochbb}
This package provides a collection of classes as functions to assemble and analyze a complex system
of random variables. To this end, it allows to analyze simple stochastic processes by means of
numerical approximation rather than stochastic simulation.
}
\author{
\packageAuthor{stochbb}

Maintainer: \packageMaintainer{stochbb}
}
\keyword{ package }
