\name{dervied}
\alias{derived}
\title{
Derived random variables.
}
\description{
In contrast to \link{atomic} random variables, derived random variables (RVs) are derived from one
or more other random variables. They may be simple transformations of RVs like \code{\link{affine}}
transformed RVs or more or less complex combinations of RVs like \code{\link{chain}} or
\code{\link{condchain}}. Like all RVs, derived RVs are constructed using factory functions listed
below.
}
\seealso{
 \code{\link{affine}},
 \code{\link{chain}},
 \code{\link{minimum}},
 \code{\link{maximum}},
 \code{\link{mixture}},
 \code{\link{conditional}},
 \code{\link{condchain}},
 \code{\link{compound}}.
}
